Moved the grid down just a tad
The grid was originally on the xz plane, and so were many of the items
being drawn in the scene. This caused a lot of z-fighting. So, to get
around this while still having a useful grid, the grid was lowered a
small amount to prevent z-fighting with the grid. This makes the demo
look much nicer.