Fix pikmin sorting
This is why we were using the enum class in the first place. Hopefully
this is somewhat more clear; we're utilizing the int-nature of pikmin
types to do a rollover check here. I added a kNone type at some point
and broke the straight integer version; this should be a bit more sane
and resiliant. Unless we add more pikmin types? (Not planning on it.)