Add beginning of blender to .fbx export script.
The only part of asset conversion that is manual at the moment is
exporting to .fbx files from within Blender. This script aims to
eventually make this step automatic, so that the entire asset pipeline
is automatic and can be run as part of the build.

Eventually, this may import the dsgx converter as a library and
synthesize the .dsgx file directly from a .blend file, removing our
reliance on the FBX SDK and improving build times.