Flesh out .fbx exporter.
With this script, the FBX exporter that uses blender as its runtime is
usable. It can select objects within the scene and export them in a
format that is suitable for the dsgx converter. It correctly handles
full scene exporting and individual object selection. It can also
detect when it is not being run within blender, and will run itself in
blender, making it easier to use.