Patch Blender FBX exporter to remove false errors
The FBX exporter in Blender 2.74 attempts to create dupli lists for all
objects that are going to be exported, even if they don't have dupli.
Since none of our objects are using this feature, this error shows up a
lot.

A patch based on
https://developer.blender.org/rBA94abe3b97f5c0b1afaa084dc8918464d27689b8
8 has been added to the build environment. Linked is the official
patch, which will make it into Blender 2.76.