Major Change: convert models directly from .blend files
This involved modifying the dsgx-converter to support
multiple meshes, then creating a blender-specific
import script to interface with blender in headless
mode. The result is a much smoother workflow; now one
can simply save a .blend file, rebuild the project,
and see the results ingame immediately.

Notably, pikmin and their onions are now multi-mesh,
sharing a single armature and animation set. When dealing
with multi-mesh actors, remember to call entity->set_mesh
to select which one to display. Single-mesh actors can
ignore this; the first mesh encountered will be used
as the default.