Implement a nearest neighbors system for physics (200k)
The basic idea is, each pikmin / object keeps a list of the closest
objects that are nearby to it. This list is very slowly (one per frame)
compared against the rest of the objects in the scene, and updated as
the objects move around relative to each other. When it's in perfect
sync, it represents the 4 closest neighboring objects.

Each collision check considers ONLY this neighboring objects list.
This means that collision is much faster (each object needs to check
its neighbor list as opposed to the entire scene) but can potentially
break more often if many objects are in close quarters. The number
of neighbors to track is variable in project_settings.h; increasing
this number makes physics more accurate in crowded scenarios, but
heavily decreases performance.

Right now, only regular objects are valid neighbors; pikmin do not
count as neighbors. (But they *track* neighbors, so the collision
works for them still.) This greatly reduces the number of objects
to consider as neighbors, and contributes to the accuracy of the
lists over time.