Support animation_mode, to select between bone and vertex animation
Note: Olimar is rendered as a T-pose in this build. This is not a
bug, but a lack of implementation: he is flagged for vertex animation,
does not ever have BONE data generated, and the engine doesn't know
how to read his ANIM data yet. This is my next task.