Implement Vertex/Normal animation in DSGX files
This should be a complete implementation of the standard, though
the exporter currently has bone transforms turned off. I've only
applied this to the Olimar model, but he animates decently. There
are tradeoffs to vertex animation versus bone animation, so we'll
need to be able to use a mix of the two going forward.