Setup fog registers, switch to W-buffering for depth
I'm really perplexed as to why W-buffering is more accurate. I
think the clip-friendly perspective matrix might be to blame.
Fog technically works but has issues with pass divisions on
hardware, but I want to leave the setup code there in case we
successfully switch to a side-to-side rendering strategy later.