Make DebugToggles a bit more sane. Uses a map in MultipassEngine.
I found the global nature of the flags and the boolean pointers to
be really unclean and generally a pain to work with. This is
probably slightly slower, but it's quite a bit simpler to deal
with all around, and easier to understand the implications of
things like deallocating various game structures.