Implement treasure detection disabling when full / retrieved
This brings behavior back in line with the pre-Handle rewrite, and is a
bit more straightforward all around. Pikmin detect when the targeting
circle they're actively chasing vanishes, and cease following it. The
treasure simply deletes its detection radius when full / retrieved, and
Pikmin proceed to cease chasing after it.

Because the detection radius is actually removed, this frees up a
neighbor slot in the kNN graph, and is slightly more efficient than the
previous implementation.