Implement a simplistic debug logging system using std::deque
The idea here is to funnel all debug messages through the logger,
which can then pass those messages around to various systems.
This should allow all messages to be (a) displayed to the in-game
console, (b) passed through nocashMessage for easier viewing in
no$gba and desmume, and optionally later (c) be written out to
the SD card for offline viewing, useful when debugging crashes
that prevent the video display from being used.