Move the debug profiler out of the multipass renderer
It's now a namespaced 'global' which is fine, as it only ever
needs to exist once for NDS projects. Its use of the hardware
timers is exclusive, so it couldn't logically exist twice anyway
without some serious coordination, and it would make more sense
to simply create additional topics as needed in that scenario.