Adjust Makefile dependencies to be parallel build friendly
While this isn't the default for maximum stability, the project can
now be built using "./build.sh make -j N", where N is the number of
cores you want to use. This can drastically reduce the compile time,
and is recommended for development.

The Makefile for arm9 already works in two sections, but having the
bin2o processing occur at the same time as code generation led to
a race condition before. Since the header files generated by bin2o
aren't recognized properly as dependencies of the code files that
use them, make would occasionally miss the dependency and fail. I
chose to solve this by moving all bin2o generation into the first
pass, rewriting the rules as appropriate. This causes all data
processing and NitroFS generation to occur before any C++ code is
compiled, ensuring that any include files are generated in advance
of any code generation. There's probably a more elegant way to set
this up, but this works reasonably well and is simple to understand.