Add a level exporter!
This uses a .blend scene as its input. Each normal rendered mesh is turned in to an appropriate Static object spawn. Each object in the scene with the "spawn" property set can spawn that type of object in. This allows using Blender as a simple level editor without needing to do anything terribly fancy; just create empty objects, set the spawn property, and place them where you want the object to end up in-game.